build: Большая реконфигурация проекта
