+++ demos/desktop_test/build.gradle.kts
@@ -1,0 +1,38 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+
+plugins {
+    kotlin("jvm")
+    id("org.jetbrains.compose")
+}
+
+group = "org.leftbrained.multimark"
+version = "1.0-SNAPSHOT"
+
+repositories {
+    mavenCentral()
+    mavenLocal()
+    maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
+    google()
+}
+
+dependencies {
+    // Note, if you develop a library, you should use compose.desktop.common.
+    // compose.desktop.currentOs should be used in launcher-sourceSet
+    // (in a separate module for demo project and in testMain).
+    // With compose.desktop.common you will also lose @Preview functionality
+    implementation(compose.desktop.currentOs)
+
+    implementation("com.leftbrained.multimark:Multimark:0.0.1")
+}
+
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "desktop_test"
+            packageVersion = "1.0.0"
+        }
+    }
+}
