+++ demos/desktop_test/gradle.properties
@@ -1,0 +1,6 @@
+org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
+kotlin.code.style=official
+kotlin.version=1.9.10
+compose.version=1.5.2
+#Compose
+org.jetbrains.compose.experimental.uikit.enabled=true
