--- desktopApp/build.gradle.kts
@@ -1,30 +1,0 @@
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-}
-
-kotlin {
-    jvm()
-    sourceSets {
-        val jvmMain by getting {
-            dependencies {
-                implementation(compose.desktop.currentOs)
-                implementation(project(":shared"))
-            }
-        }
-    }
-}
-
-compose.desktop {
-    application {
-        mainClass = "MainKt"
-
-        nativeDistributions {
-            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
-            packageName = "KotlinMultiplatformComposeDesktopApplication"
-            packageVersion = "1.0.0"
-        }
-    }
-}
