--- desktopApp/src/jvmMain/kotlin/main.kt
@@ -1,9 +1,0 @@
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(onCloseRequest = ::exitApplication, title = "Multimark", icon = painterResource("logo.svg")) {
-        MainView()
-    }
-}
