+++ gradle.properties
@@ -8,6 +8,7 @@ kotlin.mpp.enableCInteropCommonization=true
 kotlin.mpp.androidSourceSetLayoutVersion=2
 #Compose
 org.jetbrains.compose.experimental.uikit.enabled=true
+org.jetbrains.compose.experimental.jscanvas.enabled=true
 #Android
 android.useAndroidX=true
 android.compileSdk=34
@@ -15,5 +16,6 @@ android.targetSdk=34
 android.minSdk=24
 #Versions
 kotlin.version=1.9.0
-agp.version=8.0.2
 compose.version=1.5.0
+
+multimark-version=0.0.1
