+++ settings.gradle.kts
@@ -1,33 +1,22 @@
 rootProject.name = "Multimark"
 
 
-include(":shared")
-include(":desktopApp")
+// include(":shared")
+// include(":desktopApp")
+include(":desktop_test")
+
+project(":desktop_test").apply {
+    projectDir = file("demos/desktop_test")
+    name = "desktop_test"
+}
 
 pluginManagement {
     repositories {
         gradlePluginPortal()
         mavenCentral()
+        mavenLocal()
         google()
         maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
-    }
 
-    plugins {
-        val kotlinVersion = extra["kotlin.version"] as String
-        val composeVersion = extra["compose.version"] as String
-
-        kotlin("jvm").version(kotlinVersion)
-        kotlin("multiplatform").version(kotlinVersion)
-
-
-        id("org.jetbrains.compose").version(composeVersion)
-    }
-}
-
-dependencyResolutionManagement {
-    repositories {
-        mavenCentral()
-        google()
-        maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
     }
 }
